requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
streamlit==1.31.1
langchain==0.1.6
langchain-google-genai==0.0.9
langchain-community==0.0.20
python-dotenv==1.0.1
redis==5.0.1
lxml==5.1.0
